% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickMapFunction.R
\name{pal}
\alias{pal}
\title{Make a choropleth map with quantile bins!
This function reads a OGR object, a selected column from that data frame to be mapped and a label for areal units.
The data is then mapped via leaflet!}
\usage{
pal(x)
}
\arguments{
\item{df}{rgdal, readOGR obejcet}

\item{column}{the data you would like to use for the choropleth map with quantile bins!}

\item{labelcolumn}{unique identifier for areal units}
}
\value{
leaflet choropleth map with quantile bins
}
\description{
Make a choropleth map with quantile bins!
This function reads a OGR object, a selected column from that data frame to be mapped and a label for areal units.
The data is then mapped via leaflet!
}
